util: move weaponType
util: move weaponType

WeaponType has been moved so it can be utilised by other plugins too. it has also been updated to reflect newer and missing weapons
